nightcycle improvements
It takes ages to iterate over all maps so we trim the number of day parts to 4. We also set the shift time start right before the top day part so it starts cycling immediately (so we're not in darkness for 20 minutes) and sort z-levels by client population so the levels with players get cycled first.