migrate teleporter triggers to named teleporters
teleporter triggers were impossible to reason about because their
destinations were varedited instead of something you could see on a map.
more importantly, their z-coordinate relied on maps loading in a
specific order and with zero changes to that order. even more
importantly than that, TG's dynamic map loader centers smaller maps if
the maxx/y is changed, making the x- and y- coordinates useless as well.
with these changes in places, we can go back to TG's dynamic map
loading.

the change was automated so the teleport destination names are pretty
generic, but they can be renamed if anyone cares enough.